Ref: Entities as render objects
